requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.3
apache-airflow-core==3.1.3
apache-airflow-providers-common-compat==1.9.0
apache-airflow-providers-common-io==1.6.5
apache-airflow-providers-common-sql==1.29.0
apache-airflow-providers-smtp==2.3.2
apache-airflow-providers-standard==1.9.2
apache-airflow-task-sdk==1.1.3
argcomplete==3.6.3
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
Automat==20.2.0
babel==2.17.0
backports.strenum==1.3.1
bcrypt==3.2.0
beautifulsoup4==4.14.2
blinker==1.4
cadwyn==5.4.5
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.3.1
cloud-init==25.2
colorama==0.4.4
colorlog==6.10.1
comm==0.2.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
dbus-python==1.2.18
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.117.1
fastapi-cli==0.0.16
filelock==3.20.0
findspark==2.0.1
fsspec==2025.10.0
gdown==5.2.0
geopandas==1.1.1
googleapis-common-protos==1.72.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.5
hyperlink==21.0.0
idna==3.3
importlib_metadata==8.7.0
incremental==21.3.0
ipykernel==7.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.2.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.19.0
narwhals==2.12.0
natsort==8.4.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
numpy==2.2.6
oauthlib==3.2.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.6.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==22.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.23
pycurl==7.44.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.42.1
pygtrie==2.5.0
PyHamcrest==2.0.2
PyJWT==2.10.1
pyogrio==0.11.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproj==3.7.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pyspark==4.0.1
python-apt==2.4.0+ubuntu4
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2022.1
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.16.0
rpds-py==0.29.0
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
service-identity==18.1.0
setproctitle==1.3.7
shapely==2.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==22.1.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.1
unattended-upgrades==0.1
universal_pathlib==0.2.6
urllib3==1.26.5
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
wadllib==1.3.6
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wirerope==1.0.0
wrapt==2.0.1
zipp==3.23.0
zope.interface==5.4.0
